+++ mitra_models/migrations/schema.sql
@@ -76,6 +76,7 @@ CREATE TABLE user_account (
     invite_code VARCHAR(100) UNIQUE REFERENCES user_invite_code (code) ON DELETE SET NULL,
     user_role SMALLINT NOT NULL,
     client_config JSONB NOT NULL DEFAULT '{}',
+    shared_client_config JSONB NOT NULL DEFAULT '{}',
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now()
 );